Fix summarizer tests
